# swagger.yaml
swagger: '2.0'
info:
  description: Api Documentation
  version: '1.0'
  title: Api Documentation
  termsOfService: urn:tos
  contact: {}
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0
host: localhost:8080
basePath: /
tags:
  - name: auth-controller
    description: Auth Controller
  - name: loan-controller
    description: Loan Controller
paths:
  /loan:
    get:
      tags:
        - loan-controller
      summary: getLoan
      operationId: getLoanUsingGET
      produces:
        - '*/*'
      parameters:
        - name: accountNonExpired
          in: query
          required: false
          type: boolean
        - name: accountNonLocked
          in: query
          required: false
          type: boolean
        - name: authorities[0].authority
          in: query
          required: false
          type: string
        - name: client.address
          in: query
          required: false
          type: string
        - name: client.firstName
          in: query
          required: false
          type: string
        - name: client.id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.iin
          in: query
          required: false
          type: string
        - name: client.lastName
          in: query
          required: false
          type: string
        - name: client.loans[0].endDate
          in: query
          required: false
          type: string
          format: date
        - name: client.loans[0].id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.loans[0].interestRate
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].loanSum
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].monthlyPay
          in: query
          required: false
          type: number
          format: double
        - name: client.loans[0].period
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].startDate
          in: query
          required: false
          type: string
          format: date
        - name: client.password
          in: query
          required: false
          type: string
        - name: client.patronymic
          in: query
          required: false
          type: string
        - name: client.phoneNumber
          in: query
          required: false
          type: string
        - name: client.roles[0].id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.roles[0].roleName
          in: query
          required: false
          type: string
        - name: credentialsNonExpired
          in: query
          required: false
          type: boolean
        - name: enabled
          in: query
          required: false
          type: boolean
        - name: password
          in: query
          required: false
          type: string
        - name: username
          in: query
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/RemainingLoan'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
    post:
      tags:
        - loan-controller
      summary: addLoan
      operationId: addLoanUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - name: accountNonExpired
          in: query
          required: false
          type: boolean
        - name: accountNonLocked
          in: query
          required: false
          type: boolean
        - name: authorities[0].authority
          in: query
          required: false
          type: string
        - name: client.address
          in: query
          required: false
          type: string
        - name: client.firstName
          in: query
          required: false
          type: string
        - name: client.id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.iin
          in: query
          required: false
          type: string
        - name: client.lastName
          in: query
          required: false
          type: string
        - name: client.loans[0].endDate
          in: query
          required: false
          type: string
          format: date
        - name: client.loans[0].id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.loans[0].interestRate
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].loanSum
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].monthlyPay
          in: query
          required: false
          type: number
          format: double
        - name: client.loans[0].period
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].startDate
          in: query
          required: false
          type: string
          format: date
        - name: client.password
          in: query
          required: false
          type: string
        - name: client.patronymic
          in: query
          required: false
          type: string
        - name: client.phoneNumber
          in: query
          required: false
          type: string
        - name: client.roles[0].id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.roles[0].roleName
          in: query
          required: false
          type: string
        - name: credentialsNonExpired
          in: query
          required: false
          type: boolean
        - name: enabled
          in: query
          required: false
          type: boolean
        - in: body
          name: loan
          description: loan
          required: true
          schema:
            $ref: '#/definitions/Loan'
        - name: password
          in: query
          required: false
          type: string
        - name: username
          in: query
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Loan'
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /loan/{id}:
    get:
      tags:
        - loan-controller
      summary: getClientLoan
      operationId: getClientLoanUsingGET
      produces:
        - '*/*'
      parameters:
        - name: accountNonExpired
          in: query
          required: false
          type: boolean
        - name: accountNonLocked
          in: query
          required: false
          type: boolean
        - name: authorities[0].authority
          in: query
          required: false
          type: string
        - name: client.address
          in: query
          required: false
          type: string
        - name: client.firstName
          in: query
          required: false
          type: string
        - name: client.id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.iin
          in: query
          required: false
          type: string
        - name: client.lastName
          in: query
          required: false
          type: string
        - name: client.loans[0].endDate
          in: query
          required: false
          type: string
          format: date
        - name: client.loans[0].id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.loans[0].interestRate
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].loanSum
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].monthlyPay
          in: query
          required: false
          type: number
          format: double
        - name: client.loans[0].period
          in: query
          required: false
          type: integer
          format: int32
        - name: client.loans[0].startDate
          in: query
          required: false
          type: string
          format: date
        - name: client.password
          in: query
          required: false
          type: string
        - name: client.patronymic
          in: query
          required: false
          type: string
        - name: client.phoneNumber
          in: query
          required: false
          type: string
        - name: client.roles[0].id
          in: query
          required: false
          type: integer
          format: int64
        - name: client.roles[0].roleName
          in: query
          required: false
          type: string
        - name: credentialsNonExpired
          in: query
          required: false
          type: boolean
        - name: enabled
          in: query
          required: false
          type: boolean
        - name: id
          in: path
          description: id
          required: true
          type: integer
          format: int64
        - name: password
          in: query
          required: false
          type: string
        - name: username
          in: query
          required: false
          type: string
      responses:
        '200':
          description: OK
          schema:
            type: array
            items:
              $ref: '#/definitions/RemainingLoan'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /login:
    post:
      tags:
        - auth-controller
      summary: login
      operationId: loginUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: authenticationDto
          description: authenticationDto
          required: true
          schema:
            $ref: '#/definitions/AuthenticationDto'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
  /registration:
    post:
      tags:
        - auth-controller
      summary: registration
      operationId: registrationUsingPOST
      consumes:
        - application/json
      produces:
        - '*/*'
      parameters:
        - in: body
          name: client
          description: client
          required: true
          schema:
            $ref: '#/definitions/Client'
      responses:
        '200':
          description: OK
          schema:
            type: string
        '201':
          description: Created
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not Found
definitions:
  AuthenticationDto:
    type: object
    properties:
      password:
        type: string
      phoneNumber:
        type: string
    title: AuthenticationDto
  Client:
    type: object
    properties:
      address:
        type: string
      firstName:
        type: string
      iin:
        type: string
      lastName:
        type: string
      password:
        type: string
      patronymic:
        type: string
      phoneNumber:
        type: string
    title: Client
  Loan:
    type: object
    properties:
      interestRate:
        type: integer
        format: int32
      loanSum:
        type: integer
        format: int32
      period:
        type: integer
        format: int32
    title: Loan
  RemainingLoan:
    type: object
    properties:
      earlyPayment:
        type: integer
        format: int32
      nextPaymentAmount:
        type: number
        format: double
      nextPaymentDate:
        type: string
        format: date
      remainingPaymentsCount:
        type: integer
        format: int64
      remains:
        type: number
        format: double
    title: RemainingLoan